Move parser specific tests utils to parser tests
